Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     char ch;
-    scanf("%f%f",&a,&b);
+    scanf("%d%d",&a,&b);
     scanf("%c",&ch);
     switch(ch){
         case '+':
-            printf("%f",a+b);
+            printf("%d",a+b);
             break;
         case '-':
-            printf("%f",a-b);
+            printf("%d",a-b);
             break;
         case '*':
-            printf("%f",a*b);
+            printf("%d",a*b);
             break;
         case '/':
             if(b!=0){
-            printf("%.2f",a/b);
+            printf("%.2f",(float)a/b);
             }
             else{
                 printf("error");
